+++ praw/reddit.py
@@ -16,11 +16,16 @@ from prawcore import (
     UntrustedAuthenticator,
     session,
 )
+from prawcore.exceptions import BadRequest
 
 from . import models
 from .config import Config
 from .const import API_PATH, USER_AGENT_FORMAT, __version__
-from .exceptions import ClientException, MissingRequiredAttributeException
+from .exceptions import (
+    ClientException,
+    MissingRequiredAttributeException,
+    RedditAPIException,
+)
 from .objector import Objector
 
 try:
@@ -648,9 +653,26 @@ class Reddit:
             (default: None).
 
         """
-        return self._core.request(
-            method, path, data=data, files=files, params=params
-        )
+        try:
+            return self._core.request(
+                method, path, data=data, files=files, params=params
+            )
+        except BadRequest as exception:
+            try:
+                data = exception.response.json()
+            except ValueError:
+                # TODO: Remove this exception after 2020-12-31 if no one has filed a bug against it.
+                raise Exception(
+                    "Unexpected BadRequest without json body. Please file a bug: https://github.com/praw-dev/praw/issues"
+                ) from exception
+            if "fields" in data:
+                assert len(data["fields"]) == 1
+                field = data["fields"][0]
+            else:
+                field = None
+            raise RedditAPIException(
+                [data["reason"], data["explanation"], field]
+            ) from exception
 
     def submission(  # pylint: disable=invalid-name,redefined-builtin
         self, id: Optional[str] = None, url: Optional[str] = None